requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
rq==1.15.0
redis==5.0.0
numpy==1.24.3
opencv-python==4.8.0.76
torch==2.0.1
torchvision==0.15.2
transformers==4.33.1
Pillow==10.0.1
pytest==7.4.2
httpx==0.24.1
paddlepaddle==2.5.1
paddleocr==2.7.0.2
pix2tex==0.1.0
pdf2image==1.16.3
albumentations==1.3.1
sympy==1.12
matplotlib==3.8.0
openai==0.28.0
